# Performance optimizations for Feed and Authorizer, with test case 