--- a/app/lib/feed.rb
+++ b/app/lib/feed.rb
@@ -6,25 +6,13 @@ class Feed
 
   # Return a list of `FeedCard`s for the feed based on event notes.
   def event_note_cards(from_time, limit)
-    # This is a performance optimization - we check the authorization
-    # of one at a time, so that we can exit early
-    recent_event_notes = unsafe_authorization! do
-      unsafe_event_notes = EventNote
-        .includes(:student)
-        .where(is_restricted: false)
-        .where('recorded_at < ?', from_time)
-        .order(recorded_at: :desc)
-
-      authorized_event_notes = []
-      unsafe_event_notes.each do |unsafe_event_note|
-        if @authorizer.is_authorized_for_note?(unsafe_event_note)
-          authorized_event_notes << unsafe_event_note
-        end
-        break if authorized_event_notes.size >= limit
-      end
-      authorized_event_notes
-    end
-
+    recent_event_notes = EventNote
+      .where(student_id: authorized_students.map(&:id))
+      .where(is_restricted: false)
+      .where('recorded_at < ?', from_time)
+      .order(recorded_at: :desc)
+      .limit(limit)
+      .includes(student: :homeroom)
     recent_event_notes.map {|event_note| event_note_card(event_note) }
   end
 
@@ -35,24 +23,23 @@ class Feed
     limit = options[:limit] || 3
     days_back = options[:days_back] || 7
     days_ahead = options[:days_ahead] || 7
-    students = @authorizer.authorized do
-      Student.select(:id, :primary_email, :first_name, :last_name, :date_of_birth)
-        .where('extract(doy from date_of_birth) > ?', time_now.yday - days_back)
-        .where('extract(doy from date_of_birth) <= ?', time_now.yday + days_ahead)
-        .order('extract(doy from date_of_birth) DESC')
-        .limit(limit)
-    end
+    students = Student
+      .where(id: authorized_students.map(&:id))
+      .where('extract(doy from date_of_birth) > ?', time_now.yday - days_back)
+      .where('extract(doy from date_of_birth) <= ?', time_now.yday + days_ahead)
+      .order('extract(doy from date_of_birth) DESC')
+      .limit(limit)
     students.map { |student| birthday_card(student, time_now) }
   end
 
   # Returns recent discipline incidents.
   def incident_cards(time_now, limit)
-    students = @authorizer.authorized { Student.all }
     incidents = DisciplineIncident
-      .where(student_id: students.map(&:id))
+      .where(student_id: authorized_students.map(&:id))
       .where('occurred_at < ?', time_now)
       .order(occurred_at: :desc)
       .limit(limit)
+      .includes(student: :homeroom)
     incidents.map {|incident| incident_card(incident) }
   end
 
@@ -63,9 +50,9 @@ class Feed
   end
 
   private
-  # This is just to tag a block as unsafe in terms of authorization
-  def unsafe_authorization!
-    yield
+  # Performance optimization to cache across methods
+  def authorized_students
+    @authorized_students ||= @authorizer.authorized { Student.all }
   end
 
   # Create json for exactly what UI needs and return as `FeedCard`