--- a/app/lib/perf_test.rb
+++ b/app/lib/perf_test.rb
@@ -1,5 +1,41 @@
 class PerfTest
-  # Holds timing data
+  # Helper class to print summary stats like p50, p95
+  #
+  # Usage:
+  #
+  #   Reporter.new.report({
+  #     unoptimized: unoptimized_timer.report,
+  #     optimized: optimized_timer.report
+  #   })
+  #
+  # or json = JSON.parse(IO.read('runs.json'));nil
+  # Reporter.new.report(json)
+  class Reporter
+    def report(timer_reports_map)
+      timer_reports_map.each do |report_key, tuples|
+        puts
+        puts "#{report_key.upcase}, sample size: #{tuples.size}"
+        tuples.group_by {|t| t[0]}.each do |key, ts|
+          values = ts.map {|t| t[1] }
+          p50 = percentile(values, 0.50).round
+          p95 = percentile(values, 0.95).round
+          puts "  #{key.ljust(25)}\tmedian: #{p50}ms\t\tp95: #{p95}ms"
+        end
+      end
+      nil
+    end
+
+    private
+    # copied from https://stackoverflow.com/questions/11784843/calculate-95th-percentile-in-ruby
+    def percentile(values, percentile)
+      values_sorted = values.sort
+      k = (percentile*(values_sorted.length-1)+1).floor - 1
+      f = (percentile*(values_sorted.length-1)+1).modulo(1)
+      return values_sorted[k] + (f * (values_sorted[k+1] - values_sorted[k]))
+    end
+  end
+
+  # Helper class to hold timing data
   class Timer
     def initialize()
       @measurements = []
@@ -27,7 +63,7 @@ class PerfTest
     end
   end
 
-  # For testing the feed
+  # Usage for testing the feed
   def self.feed(percentage, options = {})
     timer = PerfTest.new.run(0.1) do |t, educator|
       time_now = options[:time_now] || Time.at(1521552855)
@@ -58,7 +94,7 @@ class PerfTest
     timer
   end
 
-  # Generic usage:
+  # Generic perftest usage:
   # timer = perf_test.run(0.10) do |t, educator|
   #   t.measure('whatever') { do_something_for(educator) }
   #   t.measure('foo') { do_something_else_for(educator) }