# Who is this PR for?
all educators, but HS classroom teachers were the initial focus

# What problem does this PR fix?
The home feed has latency up to 30 seconds for some HS teachers, particularly those who don't have schoolwide, gradelevel, or districtwide access.  This makes this mostly unusable beyond a prototype, and prevents shipping this as the home page for all educators.

# What does this PR do?
Makes a few performance optimizations to `Feed` and to `Authorizer`.  There's no functional changes, just optimizing the query latency for the feed endpoint.  This also adds the test case and some maybe useful code for other perf test cases.  No tests break because there are no functional changes.

The test case works by grabbing some educators, then essentially running `home_controller#feed_json` with some more instrumentation around it.  I ran this with a known set of educators I was trying to optimize for, then generalized to a consistent sample of 10% of educators.  The test runs below don't all use the same set of educators, and even when they do there's some variability across tests runs.  I did test runs in new one-off dynos to try to get repeatable results, but there's still just a bunch of variability.  But overall this cuts the p95 to ~41 seconds to ~3 seconds for total request latency (with raw data below).

This PR includes two main optimizations.  The first is to `Authorizer`, which was repeatedly querying for educator section and homeroom data.  I'm not sure why this change works (it's counter-intuitive to me), but by turning on ActiveRecord debug logging the effect is clear.  The second  is to `Feed`, which came from noticing that `event_note_cards` and `incident_cards` both were slow enough that it might be worth paralleling out in the client.  I looked at what they were doing, and rolled back https://github.com/studentinsights/studentinsights/pull/1548 to optimize across all methods with the new improvement to `Authorizer`.  This worked great because if we just cache `authorized_students` within the feed, it simplifies all the other methods and is also faster.

test case data:
```
MASTER_SMALL, sample size: 45
  all_for_educator         	median: 14381ms		p95: 41457ms
  birthday_cards           	median: 16ms		p95: 30ms
  event_note_cards         	median: 5270ms		p95: 24921ms
  feed_cards               	median: 0ms		p95: 0ms
  incident_cards           	median: 8662ms		p95: 16662ms

FIRST_CDB25D5_SMALL, sample size: 45
  all_for_educator         	median: 4270ms		p95: 4917ms
  birthday_cards           	median: 12ms		p95: 48ms
  event_note_cards         	median: 2493ms		p95: 2640ms
  feed_cards               	median: 0ms		p95: 0ms
  incident_cards           	median: 1800ms		p95: 2335ms

FIRST_CDB25D5_TEN, sample size: 645
  all_for_educator         	median: 3588ms		p95: 5283ms
  birthday_cards           	median: 11ms		p95: 16ms
  event_note_cards         	median: 2008ms		p95: 2778ms
  feed_cards               	median: 0ms		p95: 0ms
  incident_cards           	median: 1530ms		p95: 2646ms

SECOND_121D502_KNOWN, sample size: 185
  all_for_educator         	median: 1652ms		p95: 2689ms
  birthday_cards           	median: 21ms		p95: 214ms
  event_note_cards         	median: 1503ms		p95: 2527ms
  feed_cards               	median: 0ms		p95: 0ms
  incident_cards           	median: 87ms		p95: 335ms

SECOND_121D502_TEN, sample size: 645
  all_for_educator         	median: 470ms		p95: 1274ms
  birthday_cards           	median: 4ms		p95: 42ms
  event_note_cards         	median: 391ms		p95: 1234ms
  feed_cards               	median: 0ms		p95: 0ms
  incident_cards           	median: 20ms		p95: 83ms

FINAL_131D502_TEN, sample size: 485
  all_for_educator         	median: 833ms		p95: 2965ms
  birthday_cards           	median: 7ms		p95: 70ms
  event_note_cards         	median: 774ms		p95: 2629ms
  feed_cards               	median: 0ms		p95: 0ms
  incident_cards           	median: 9ms		p95: 211ms

FINALFINAL_131D502_TEN, sample size: 485
  all_for_educator         	median: 464ms		p95: 1736ms
  birthday_cards           	median: 4ms		p95: 42ms
  event_note_cards         	median: 439ms		p95: 1673ms
  feed_cards               	median: 0ms		p95: 0ms
  incident_cards           	median: 13ms		p95: 102ms
```